Changed default type of indices in IndexedTriMesh/IndexedTriangle back to uint16 as it was visualized wrongly when using 32bit.
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@2872 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/geometry/IndexedTriMesh.hpp
@@ -216,7 +216,7 @@ namespace geometry {
 	 * @brief an Indexed Triangle mesh
 	 *
 	 */
-	template <class T=double, class S=uint32_t>
+	template <class T=double, class S=uint16_t>
 	class IndexedTriMeshN0: public IndexedTriMesh<T> {
 	public:
 	    typedef T value_type;
